--- a/aten/src/ATen/native/cuda/Embedding.cu
+++ b/aten/src/ATen/native/cuda/Embedding.cu
@@ -8,6 +8,7 @@
 #include <THC/THCDeviceUtils.cuh>
 #include <THC/THCTensorMathReduce.cuh>
 #include <THC/THCThrustAllocator.cuh>
+#include <THC/THCReduce.cuh>
 
 #include <thrust/execution_policy.h>
 #include <thrust/iterator/constant_iterator.h>