--- a/aten/src/TH/generic/THTensorMoreMath.cpp
+++ b/aten/src/TH/generic/THTensorMoreMath.cpp
@@ -244,63 +244,6 @@ LAB_IMPLEMENT_BASIC_FUNCTION(abs,TH_MATH_NAME(fabs))
 LAB_IMPLEMENT_BASIC_FUNCTION(cosh,TH_MATH_NAME(cosh),HYPER_TH_OMP_OVERHEAD_THRESHOLD)
 LAB_IMPLEMENT_BASIC_FUNCTION(tanh,TH_MATH_NAME(tanh),HYPER_TH_OMP_OVERHEAD_THRESHOLD)
 
-void THTensor_(renorm)(THTensor *res, THTensor *src, scalar_t value, int dimension, scalar_t maxnorm)
-{
-  THTensor *rowR, *rowS;
-  dimension = at::maybe_wrap_dim(dimension, src);
-  THArgCheck(dimension >= 0 && dimension < THTensor_(nDimensionLegacyNoScalars)(src), 3, "invalid dimension %d",
-      dimension);
-  THArgCheck(value > 0, 2, "non-positive-norm not supported");
-  THArgCheck(THTensor_(nDimensionLegacyNoScalars)(src) > 1, 1, "need at least 2 dimensions, got %d dimensions",
-      THTensor_(nDimensionLegacyNoScalars)(src));
-
-  rowR = THTensor_(new)();
-  rowS = THTensor_(new)();
-
-  THTensor_(resizeAs)(res, src);
-
-  for (int64_t i = 0; i < THTensor_sizeLegacyNoScalars(src, dimension); i++)
-  {
-    scalar_t norm = 0;
-    scalar_t new_norm;
-
-    THTensor_(select)(rowS, src, dimension, i);
-    THTensor_(select)(rowR, res, dimension, i);
-    if (value == 1) {
-      TH_TENSOR_APPLY(scalar_t, rowS, norm += fabs(*rowS_data););
-    } else if (value == 2) {
-      TH_TENSOR_APPLY(scalar_t, rowS, accreal z = *rowS_data; norm += z*z;);
-    } else if (value == INFINITY) {
-      TH_TENSOR_APPLY(scalar_t, rowS, norm = THMax(norm, TH_MATH_NAME(fabs)(*rowS_data)););
-    } else {
-      TH_TENSOR_APPLY(scalar_t, rowS, norm += TH_MATH_NAME(pow)(TH_MATH_NAME(fabs)(*rowS_data), value););
-    }
-
-    if (value != INFINITY) {
-      norm = pow(norm, 1/value);
-    }
-
-    if (norm > maxnorm)
-    {
-      new_norm = maxnorm / (norm + 1e-7);
-
-      TH_TENSOR_APPLY2(
-        scalar_t, rowR, scalar_t, rowS,
-        *rowR_data = (*rowS_data) * new_norm;
-      )
-    }
-    else
-    {
-      at::Tensor rowR_wrap = THTensor_wrap(rowR);
-      at::Tensor rowS_wrap = THTensor_wrap(rowS);
-      at::native::copy_(rowR_wrap, rowS_wrap);
-    }
-  }
-
-  c10::raw::intrusive_ptr::decref(rowR);
-  c10::raw::intrusive_ptr::decref(rowS);
-}
-
 accreal THTensor_(std_var_all)(THTensor* tensor, int64_t correction, bool take_sqrt)
     __ubsan_ignore_float_divide_by_zero__ {
   accreal mean = THTensor_wrap(tensor).mean().item<accreal>();