--- a/aten/src/THC/THCTensorMath.h
+++ b/aten/src/THC/THCTensorMath.h
@@ -28,15 +28,6 @@
 #include <THC/generic/THCTensorMathPointwise.h>
 #include <THC/THCGenerateBoolType.h>
 
-#include <THC/generic/THCTensorMathReduce.h>
-#include <THC/THCGenerateAllTypes.h>
-
-#include <THC/generic/THCTensorMathReduce.h>
-#include <THC/THCGenerateBoolType.h>
-
-#include <THC/generic/THCTensorMathReduce.h>
-#include <THC/THCGenerateBFloat16Type.h>
-
 #include <THC/generic/THCTensorScatterGather.h>
 #include <THC/THCGenerateAllTypes.h>