--- a/aten/src/THC/generated/THCTensorMathReduceBFloat16.cu
+++ b/aten/src/THC/generated/THCTensorMathReduceBFloat16.cu
@@ -1,5 +1,0 @@
-#include <THC/THCTensorMathReduce.cuh>
-#include <THC/THCTensor.hpp>
-
-#include <THC/generic/THCTensorMathReduce.cu>
-#include <THC/THCGenerateBFloat16Type.h>