--- a/aten/src/THC/generated/THCTensorMathReduceChar.cu
+++ b/aten/src/THC/generated/THCTensorMathReduceChar.cu
@@ -1,5 +1,0 @@
-#include <THC/THCTensorMathReduce.cuh>
-#include <THC/THCTensor.hpp>
-
-#include <THC/generic/THCTensorMathReduce.cu>
-#include <THC/THCGenerateCharType.h>