--- a/aten/src/THC/generated/THCTensorMathReduceLong.cu
+++ b/aten/src/THC/generated/THCTensorMathReduceLong.cu
@@ -1,5 +1,0 @@
-#include <THC/THCTensorMathReduce.cuh>
-#include <THC/THCTensor.hpp>
-
-#include <THC/generic/THCTensorMathReduce.cu>
-#include <THC/THCGenerateLongType.h>