… BUILD.bazel
… aten/src/ATen/LegacyTHFunctionsCPU.cpp
… aten/src/ATen/cuda/LegacyTHFunctionsCUDA.cpp
… aten/src/THC/generic/THCTensorMathReduce.cu
… aten/src/THC/generic/THCTensorMathReduce.h