@ngimel 2021-05-28
@peterbell10 I fixed vector_norm to not do explicit casting in #59134, so after it's merged hopefully perf will be back.