@ngimel 2021-05-30
@peterbell10 internal builds are failing with
```
stderr: ld.lld: error: undefined symbol: at::native::DispatchStub<void (*)(at::TensorIteratorBase&, double), at::native::renorm_scale_factor_stub>::DEFAULT
>>> referenced by SmallVector.h:0 (buck-out/gen/fe3a39b8/xplat/caffe2/c10/c10Android#header-mode-symlink-tree-with-header-map,headers/c10/util/SmallVector.h:0)
>>>               buck-out/gen/fe3a39b8/fbandroid/apps/fb4a/fastparse/flavors/fbandroid_debug_pt_code_gen_libAndroid#android-armv7,compile-pic-Normalization.cpp.oc5b093a4/aten/src/ATen/native/Normalization.cpp.o:(at::native::structured_renorm_out::impl(at::Tensor const&, c10::Scalar const&, long long, c10::Scalar const&, at::Tensor const&))
clang++: error: linker command failed with exit code 1 (use -v to see invocation)
```