@peterbell10 2021-05-31
I don't know much about how the mobile builds work. Is it possible that  `RenormKernel.cpp` just needs to be added to the build definition?