@ngimel 2021-06-01
I've tried adding it to tools/build_variables.bzl, let's see if it fixes the failures.